refactoring to avoid schemas in model
