+++ met-api/src/met_api/services/comment_service.py
@@ -3,7 +3,7 @@
 from met_api.constants.comment_status import Status
 from met_api.models.comment import Comment
 from met_api.models.comment_status import CommentStatus
-from met_api.models.data_class import PaginationOptions
+from met_api.models.pagination_options import PaginationOptions
 from met_api.schemas.comment import CommentSchema
 from met_api.schemas.submission import SubmissionSchema
 from met_api.schemas.survey import SurveySchema
