+++ met-api/src/met_api/services/survey_service.py
@@ -1,7 +1,7 @@
 """Service for survey management."""
 
 from met_api.constants.engagement_status import Status
-from met_api.models.data_class import PaginationOptions
+from met_api.models.pagination_options import PaginationOptions
 from met_api.models import Engagement as EngagementModel
 from met_api.models import Survey as SurveyModel
 from met_api.schemas.engagement import EngagementSchema
